backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.1.6
django-allauth==65.4.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
django-oauth-toolkit==3.0.1
django-rest-framework-social-oauth2==1.2.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
ecdsa==0.19.0
idna==3.10
jwcrypto==1.5.6
oauthlib==3.2.2
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
PyJWT==2.10.1
python-jose==3.4.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
